Moved ProviderFavicons to assets, added a rule to copy them in the built package
