+++ libs/ui/packages/react/src/assets/providersFavicons/Aave.png
