+++ libs/ui/packages/react/src/assets/providersFavicons/Baanx.png
