+++ libs/ui/packages/react/src/assets/providersFavicons/Banxa.png
