+++ libs/ui/packages/react/src/assets/providersFavicons/Bitrefill.png
