+++ libs/ui/packages/react/src/assets/providersFavicons/BtcDirect.png
