+++ libs/ui/packages/react/src/assets/providersFavicons/Changelly.png
