+++ libs/ui/packages/react/src/assets/providersFavicons/CoinGecko.png
