+++ libs/ui/packages/react/src/assets/providersFavicons/Coinify.png
