+++ libs/ui/packages/react/src/assets/providersFavicons/Compound.png
