+++ libs/ui/packages/react/src/assets/providersFavicons/Deversifi.png
