+++ libs/ui/packages/react/src/assets/providersFavicons/Ftx.png
