+++ libs/ui/packages/react/src/assets/providersFavicons/Ftxus.png
