+++ libs/ui/packages/react/src/assets/providersFavicons/Kiln.png
