+++ libs/ui/packages/react/src/assets/providersFavicons/LedgerBlack.png
