+++ libs/ui/packages/react/src/assets/providersFavicons/LedgerCard.png
