+++ libs/ui/packages/react/src/assets/providersFavicons/LedgerWhite.png
