+++ libs/ui/packages/react/src/assets/providersFavicons/Lido.png
