+++ libs/ui/packages/react/src/assets/providersFavicons/LoopiPay.png
