+++ libs/ui/packages/react/src/assets/providersFavicons/Mercuryo.png
