+++ libs/ui/packages/react/src/assets/providersFavicons/Moonpay.png
