+++ libs/ui/packages/react/src/assets/providersFavicons/Opensea.png
