+++ libs/ui/packages/react/src/assets/providersFavicons/PancakeSwap.png
