+++ libs/ui/packages/react/src/assets/providersFavicons/Paraswap.png
