+++ libs/ui/packages/react/src/assets/providersFavicons/Poap.png
