+++ libs/ui/packages/react/src/assets/providersFavicons/RainbowMe.png
