+++ libs/ui/packages/react/src/assets/providersFavicons/Ramp.png
