+++ libs/ui/packages/react/src/assets/providersFavicons/Rarible.png
