+++ libs/ui/packages/react/src/assets/providersFavicons/Simplex.png
