+++ libs/ui/packages/react/src/assets/providersFavicons/SushiSwap.png
