+++ libs/ui/packages/react/src/assets/providersFavicons/Transak.png
