+++ libs/ui/packages/react/src/assets/providersFavicons/Wyre.png
