+++ libs/ui/packages/react/src/assets/providersFavicons/Zerion.png
