+++ libs/ui/packages/react/src/assets/providersFavicons/_1Inch.png
