+++ libs/ui/packages/react/src/assets/providersFavicons/index.d.ts
